refactor(web): clean up admin ClientLayout and move to layouts/admin/
- Delete SETTINGS_LAYOUT_PREFIXES list and hasOwnLayout padding logic (all
  admin pages now manage their own layout via SettingsLayouts)
- Remove stale /admin/embeddings from hasCustomSidebar (route doesn't exist)
- Fix payment reminder banner: drop dark: modifiers, replace raw Tailwind
  colors with design-system tokens, use Opal Text + markdown()
- Remove enableCloud SSR prop threading; AdminSidebar now reads
  NEXT_PUBLIC_CLOUD_ENABLED directly from constants
- Move ClientLayout and Layout from components/admin/ to layouts/admin/
